고객 그룹 수정1
